added validation, checks and notifications

+++ src/components/dashboard/DashboardEditAndDelete/DashboardEditAndDelete.jsx
@@ -1,4 +1,5 @@
 import Link from 'next/link';
+import { toast } from 'react-toastify';
 import { handleDeleteImgFromCloudinary } from '@/utils/handleDeleteImgFromCloudinary';
 import { handleDeleteCardFromDB } from '@/utils/handleDeleteCardFromDB';
 import { createImagesArrayForDeletingFromCloudinary } from '@/utils/createImagesArrayForDeletingFromCloudinary';
@@ -26,11 +27,15 @@ const DashboardEditAndDelete = ({ data, pathname, mutate, isOwner }) => {
             {isOwner && <svg
                 className={styles.deleteIcon}
                 onClick={() => {
-                    const arrForDeleting = createImagesArrayForDeletingFromCloudinary(data);
-                    arrForDeleting.map(item => handleDeleteImgFromCloudinary(item));
-                    handleDeleteCardFromDB(url, mutate);
-                    console.log(`Delete card of ${data.slug}`)
-                }}
+                    if (confirm("Ви впевнені, що хочете видалити цю картку?")) {
+                        const arrForDeleting = createImagesArrayForDeletingFromCloudinary(data);
+                        arrForDeleting.map(item => handleDeleteImgFromCloudinary(item));
+                        handleDeleteCardFromDB(url, mutate);
+
+                        toast.success(`Картка продукту "${data?.name}" була видалена!`);
+                    }
+                }
+                }
             >
                 <use href="/sprite.svg#icon-delete" />
             </svg>}
